// star-ui/src/proto/star/v1/ui.ts
export interface OdometryData {
    /**
     * X position in metres, world frame.
     *
     * @generated from protobuf field: double x_m = 1
     */
    xM: number;
    /**
     * Y position in metres, world frame.
     *
     * @generated from protobuf field: double y_m = 2
     */
    yM: number;
    /**
     * Heading angle in radians [-pi, pi].
     *
     * @generated from protobuf field: double theta_rad = 3
     */
    thetaRad: number;
    /**
     * Forward linear velocity in m/s.
     *
     * @generated from protobuf field: double linear_velocity_mps = 4
     */
    linearVelocityMps: number;
    /**
     * Angular velocity in rad/s.
     *
     * @generated from protobuf field: double angular_velocity_rad_per_s = 5
     */
    angularVelocityRadPerS: number;
    /**
     * Source timestamp in microseconds (CLOCK_MONOTONIC).
     *
     * @generated from protobuf field: int64 timestamp_us = 6
     */
    timestampUs: string;
}
/**
 * LidarScan is a single 360deg scan from the RPLiDAR C1.
 * angle_rad, range_m, and intensity arrays are parallel; all have the same length.
 * Maximum samples per revolution: 500 (RPLiDAR C1 @ 10 Hz: 5 kHz sample rate / 10 Hz = 500 samples/rev).
 * nanopb static bound: max_count:500 for all three arrays (see ui.options).
 *
 * @generated from protobuf message star.v1.LidarScan
 */
export interface LidarScan {
    /**
     * Angle in radians for each sample [0, 2*pi).
     * nanopb max_count:500 (RPLiDAR C1 samples/rev; see ui.options).
     *
     * @generated from protobuf field: repeated float angle_rad = 1
     */
    angleRad: number[];
    /**
     * Range in metres for each sample (0 = invalid/out-of-range).
     * nanopb max_count:500 (RPLiDAR C1 samples/rev; see ui.options).
     *
     * @generated from protobuf field: repeated float range_m = 2
     */
    rangeM: number[];
    /**
     * Reflectance intensity in arbitrary sensor units (dimensionless, sensor-specific scale).
     * nanopb max_count:500 (RPLiDAR C1 samples/rev; see ui.options).
     *
     * @generated from protobuf field: repeated float intensity = 3
     */
    intensity: number[];
    /**
     * Scan start timestamp in microseconds (CLOCK_MONOTONIC, same epoch as other telemetry).
     *
     * @generated from protobuf field: int64 timestamp_us = 4
     */
    timestampUs: string;
}
// ...
class OdometryData$Type extends MessageType<OdometryData> {
    constructor() {
        super("star.v1.OdometryData", [
            { no: 1, name: "x_m", kind: "scalar", T: 1 /*ScalarType.DOUBLE*/ },
            { no: 2, name: "y_m", kind: "scalar", T: 1 /*ScalarType.DOUBLE*/ },
            { no: 3, name: "theta_rad", kind: "scalar", T: 1 /*ScalarType.DOUBLE*/ },
            { no: 4, name: "linear_velocity_mps", kind: "scalar", T: 1 /*ScalarType.DOUBLE*/ },
            { no: 5, name: "angular_velocity_rad_per_s", kind: "scalar", T: 1 /*ScalarType.DOUBLE*/ },
            { no: 6, name: "timestamp_us", kind: "scalar", T: 3 /*ScalarType.INT64*/ }
        ]);
    }
    create(value?: PartialMessage<OdometryData>): OdometryData {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.xM = 0;
        message.yM = 0;
        message.thetaRad = 0;
        message.linearVelocityMps = 0;
        message.angularVelocityRadPerS = 0;
        message.timestampUs = "0";
        if (value !== undefined)
            reflectionMergePartial<OdometryData>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: OdometryData): OdometryData {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* double x_m */ 1:
                    message.xM = reader.double();
                    break;
                case /* double y_m */ 2:
                    message.yM = reader.double();
                    break;
                case /* double theta_rad */ 3:
                    message.thetaRad = reader.double();
                    break;
                case /* double linear_velocity_mps */ 4:
                    message.linearVelocityMps = reader.double();
                    break;
                case /* double angular_velocity_rad_per_s */ 5:
                    message.angularVelocityRadPerS = reader.double();
                    break;
                case /* int64 timestamp_us */ 6:
                    message.timestampUs = reader.int64().toString();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: OdometryData, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* double x_m = 1; */
        if (message.xM !== 0)
            writer.tag(1, WireType.Bit64).double(message.xM);
        /* double y_m = 2; */
        if (message.yM !== 0)
            writer.tag(2, WireType.Bit64).double(message.yM);
        /* double theta_rad = 3; */
        if (message.thetaRad !== 0)
            writer.tag(3, WireType.Bit64).double(message.thetaRad);
        /* double linear_velocity_mps = 4; */
        if (message.linearVelocityMps !== 0)
            writer.tag(4, WireType.Bit64).double(message.linearVelocityMps);
        /* double angular_velocity_rad_per_s = 5; */
        if (message.angularVelocityRadPerS !== 0)
            writer.tag(5, WireType.Bit64).double(message.angularVelocityRadPerS);
        /* int64 timestamp_us = 6; */
        if (message.timestampUs !== "0")
            writer.tag(6, WireType.Varint).int64(message.timestampUs);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message star.v1.OdometryData
 */
export const OdometryData = new OdometryData$Type();
// @generated message type with reflection information, may provide speed optimized methods
class LidarScan$Type extends MessageType<LidarScan> {
    constructor() {
        super("star.v1.LidarScan", [
            { no: 1, name: "angle_rad", kind: "scalar", repeat: 1 /*RepeatType.PACKED*/, T: 2 /*ScalarType.FLOAT*/ },
            { no: 2, name: "range_m", kind: "scalar", repeat: 1 /*RepeatType.PACKED*/, T: 2 /*ScalarType.FLOAT*/ },
            { no: 3, name: "intensity", kind: "scalar", repeat: 1 /*RepeatType.PACKED*/, T: 2 /*ScalarType.FLOAT*/ },
            { no: 4, name: "timestamp_us", kind: "scalar", T: 3 /*ScalarType.INT64*/ }
        ]);
    }
    create(value?: PartialMessage<LidarScan>): LidarScan {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.angleRad = [];
        message.rangeM = [];
        message.intensity = [];
        message.timestampUs = "0";
        if (value !== undefined)
            reflectionMergePartial<LidarScan>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: LidarScan): LidarScan {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated float angle_rad */ 1:
                    if (wireType === WireType.LengthDelimited)
                        for (let e = reader.int32() + reader.pos; reader.pos < e;)
                            message.angleRad.push(reader.float());
                    else
                        message.angleRad.push(reader.float());
                    break;
                case /* repeated float range_m */ 2:
                    if (wireType === WireType.LengthDelimited)
                        for (let e = reader.int32() + reader.pos; reader.pos < e;)
                            message.rangeM.push(reader.float());
                    else
                        message.rangeM.push(reader.float());
                    break;
                case /* repeated float intensity */ 3:
                    if (wireType === WireType.LengthDelimited)
                        for (let e = reader.int32() + reader.pos; reader.pos < e;)
                            message.intensity.push(reader.float());
                    else
                        message.intensity.push(reader.float());
                    break;
                case /* int64 timestamp_us */ 4:
                    message.timestampUs = reader.int64().toString();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: LidarScan, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated float angle_rad = 1; */
        if (message.angleRad.length) {
            writer.tag(1, WireType.LengthDelimited).fork();
            for (let i = 0; i < message.angleRad.length; i++)
                writer.float(message.angleRad[i]);
            writer.join();
        }
        /* repeated float range_m = 2; */
        if (message.rangeM.length) {
            writer.tag(2, WireType.LengthDelimited).fork();
            for (let i = 0; i < message.rangeM.length; i++)
                writer.float(message.rangeM[i]);
            writer.join();
        }
        /* repeated float intensity = 3; */
        if (message.intensity.length) {
            writer.tag(3, WireType.LengthDelimited).fork();
            for (let i = 0; i < message.intensity.length; i++)
                writer.float(message.intensity[i]);
            writer.join();
        }
        /* int64 timestamp_us = 4; */
        if (message.timestampUs !== "0")
            writer.tag(4, WireType.Varint).int64(message.timestampUs);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message star.v1.LidarScan
 */
export const LidarScan = new LidarScan$Type();